requirements.txt
biopython==1.81
typing==3.7.4.3
